requirements.txt
Django==2.2.9
django-storages==1.9.1
boto3==1.13.10
python-decouple==3.3
